+++ packages/ui/src/config/constants.ts
@@ -1,3 +1,16 @@
+/**
+ * The default AI Models for ConfigPanel component.
+ *
+ * It includes the following providers:
+ * - DeepSeek
+ * - OpenAI
+ * - Anthropic
+ * - Google
+ * - XAI
+ * - Ollama
+ *
+ * You can override the default models by providing a different models.
+ */
 export const MODEL_PROVIDERS = {
   deepseek: {
     name: 'DeepSeek',
@@ -6,23 +19,40 @@ export const MODEL_PROVIDERS = {
   openai: {
     name: 'OpenAI',
     models: [
-      'gpt-o1',
-      'gpt-o1-preview',
-      'gpt-o1-mini',
+      'gpt-4.1',
       'gpt-4o',
       'gpt-4o-mini',
+      'gpt-4',
       'gpt-3.5-turbo',
-      'gpt-3.5-turbo-0125',
+      'o1',
+      'o1-preview',
+      'o1-mini',
+    ],
+  },
+  anthropic: {
+    name: 'Anthropic',
+    models: [
+      'claude-3.7-sonnet',
+      'claude-3.5-sonnet',
+      'claude-3.5-haiku',
+      'claude-3-opus',
+      'claude-3-sonnet',
+      'claude-3-haiku',
     ],
   },
   google: {
     name: 'Google',
     models: [
+      'gemini-2.5-pro-exp-03-25',
       'gemini-2.0-flash',
       'gemini-1.5-flash',
       'gemini-1.5-pro',
-      'gemini-1.0-pro'
+      'gemini-1.0-pro',
     ],
+  },
+  xai: {
+    name: 'Grok',
+    models: ['grok-3', 'grok-3-fast', 'grok-3-mini', 'grok-3-mini-fast'],
   },
   ollama: {
     name: 'Ollama',
@@ -42,4 +72,4 @@ export const MODEL_PROVIDERS = {
       'phi-3',
     ],
   },
-}
+};